Update cdmrmmdvmprotocol.cpp
Create different DSTAR module IDs for different incoming DMR-IDs to prevent overlapping for repeaters which use the same callsign for different DSTAR, DMR, YSF etc. services.
The module ID is created from the last digit of the DMR-ID, starting with charcter "H" to reserve A-G for real DSTAR repeater modules.
The way of calculation makes sure that the ID will always be the same for a specific service. 
There is a minor risk that a repeater uses different DMR-IDs with the same final digit.
It is just a simple and quick hack with a single line of code.